refactor: split passwordUtils into separate calculatePasswordStrength and getPasswordStrengthClass folders with co-located tests
