+++ src/utils/getPasswordStrengthClass/getPasswordStrengthClass.ts
@@ -1,0 +1,2 @@
+export const getPasswordStrengthClass = (strength: string) =>
+  `strength-${strength.toLowerCase().replace(' ', '-')}`;
